# financeApp-backend-feature-admin-group-management/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Finance Backend API",
        "description": "RESTful API backend for dual-version finance management application supporting Admin and User versions of Flutter mobile app",
        "contact": {
            "email": "support@financeapp.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/v1/transfers": {
            "get": {
                "tags": [
                    "Transfers"
                ],
                "summary": "List transfers",
                "description": "Get paginated list of transfers with optional exchange data. Regular users see only their transfers, admins see all.",
                "operationId": "9260137921902e1baa83bd603eca7227",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    },
                    {
                        "name": "transfer_date_from",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "transfer_date_to",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "sync_status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pending",
                                "syncing",
                                "synced",
                                "failed"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transfers retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Transfer"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginationMeta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Transfers"
                ],
                "summary": "Create transfer",
                "description": "Creates a new transfer record with optional currency exchange information",
                "operationId": "a8c56a4df5ccaa0f908ce4be916fed02",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "recipient_name",
                                    "amount_usd",
                                    "transfer_date"
                                ],
                                "properties": {
                                    "recipient_name": {
                                        "type": "string",
                                        "example": "Jane Smith"
                                    },
                                    "amount_usd": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 100
                                    },
                                    "transfer_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-10-22"
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Monthly payment",
                                        "nullable": true
                                    },
                                    "sync_status": {
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "syncing",
                                            "synced",
                                            "failed"
                                        ],
                                        "example": "synced"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Transfer created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Transfer created successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Transfer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/v1/transfers/{id}/exchange": {
            "post": {
                "tags": [
                    "Transfers"
                ],
                "summary": "Add exchange to transfer",
                "description": "Links currency exchange data to a transfer record with exchange rate snapshots",
                "operationId": "f0af59a5a8f176e3f428451dc089f099",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Transfer ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "exchange_date"
                                ],
                                "properties": {
                                    "converted_amount_syp": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 250000,
                                        "nullable": true
                                    },
                                    "converted_amount_try": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 3000,
                                        "nullable": true
                                    },
                                    "exchange_rate_usd_to_syp": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 2500,
                                        "nullable": true
                                    },
                                    "exchange_rate_usd_to_try": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 30,
                                        "nullable": true
                                    },
                                    "exchange_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-10-22"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Exchange added successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Exchange added successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Exchange"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Transfer not found"
                    },
                    "422": {
                        "description": "Validation error or exchange already exists"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Transfer": {
                "title": "Transfer",
                "description": "Transfer model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "recipient_name": {
                        "type": "string",
                        "example": "Jane Smith"
                    },
                    "amount_usd": {
                        "type": "number",
                        "format": "float",
                        "example": 100
                    },
                    "transfer_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-10-22"
                    },
                    "notes": {
                        "type": "string",
                        "example": "Monthly payment",
                        "nullable": true
                    },
                    "sync_status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "syncing",
                            "synced",
                            "failed"
                        ],
                        "example": "synced"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "exchange": {
                        "$ref": "#/components/schemas/Exchange"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "type": "object"
            },
            "Exchange": {
                "title": "Exchange",
                "description": "Currency exchange model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "transfer_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "converted_amount_syp": {
                        "type": "number",
                        "format": "float",
                        "example": 250000,
                        "nullable": true
                    },
                    "converted_amount_try": {
                        "type": "number",
                        "format": "float",
                        "example": 3000,
                        "nullable": true
                    },
                    "exchange_rate_usd_to_syp": {
                        "type": "number",
                        "format": "float",
                        "example": 2500,
                        "nullable": true
                    },
                    "exchange_rate_usd_to_try": {
                        "type": "number",
                        "format": "float",
                        "example": 30,
                        "nullable": true
                    },
                    "exchange_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-10-22"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "PaginationMeta": {
                "title": "Pagination Meta",
                "description": "Pagination metadata",
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 5
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 15
                    },
                    "total": {
                        "type": "integer",
                        "example": 75
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token authentication",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}